Generalize TorchHook to FrameworkHook (#2561)
* create FrameworkHook and move generic hooking functionality over

* PEP-ify import statements all around

* PR comment - fix docstring
