+++ syft/frameworks/torch/torch_attributes.py
@@ -4,6 +4,7 @@ from typing import Callable
 from typing import Any
 
 from syft.frameworks.attributes import FrameworkAttributes
+from syft.frameworks.torch.tensors.interpreters.native import TorchTensor
 
 
 class TorchAttributes(FrameworkAttributes):
@@ -25,7 +26,9 @@ class TorchAttributes(FrameworkAttributes):
         hook: A TorchHook to stash
     """
 
+    ### Subclasses must provide the following class attributes ###
     ALIAS = "torch"
+    Tensor = TorchTensor
 
     def __init__(self, torch: ModuleType, hook: ModuleType) -> None:
         """Initialization of the TorchAttributes class."""
