+++ syft/frameworks/types.py
@@ -16,6 +16,7 @@ if dependency_check.torch_available:
     import torch
 
     framework_tensors.append(torch.Tensor)
+    framework_tensors.append(torch.nn.Parameter)
     framework_shapes.append(torch.Size)
 
 framework_tensors = tuple(framework_tensors)
