+++ syft/workers/websocket_server.py
@@ -1,24 +1,25 @@
+import asyncio
 import binascii
+import logging
+import socket
+import ssl
+import sys
 from typing import Union
 from typing import List
 
-import asyncio
+import tblib.pickling_support
 import torch
 import websockets
-import ssl
-import sys
-import tblib.pickling_support
-import socket
-import logging
-
-tblib.pickling_support.install()
 
 import syft as sy
+from syft.federated import FederatedClient
 from syft.generic.tensor import AbstractTensor
 from syft.workers.virtual import VirtualWorker
+
 from syft.exceptions import GetNotPermittedError
 from syft.exceptions import ResponseSignatureError
-from syft.federated import FederatedClient
+
+tblib.pickling_support.install()
 
 
 class WebsocketServerWorker(VirtualWorker, FederatedClient):
