+++ test/torch/tensors/test_native.py
@@ -1,8 +1,8 @@
 import pytest
-
 import torch
 
 from syft.generic.pointers import PointerTensor
+
 from syft.exceptions import InvalidTensorForRemoteGet
