+++ test/workers/test_virtual.py
@@ -1,18 +1,17 @@
 from time import time
 from unittest.mock import patch
 
+import pytest
+import torch
+
 import syft as sy
-from syft.exceptions import GetNotPermittedError
-from syft.workers.virtual import VirtualWorker
 from syft import messaging
 from syft import serde
 from syft import messaging
-
 from syft.generic import pointers
+from syft.workers.virtual import VirtualWorker
 
-import pytest
-import torch
-import torch as th
+from syft.exceptions import GetNotPermittedError
 
 
 def test_send_msg():
@@ -198,7 +197,7 @@ def test_obj_not_found(workers):
 
     bob = workers["bob"]
 
-    x = th.tensor([1, 2, 3, 4, 5]).send(bob)
+    x = torch.tensor([1, 2, 3, 4, 5]).send(bob)
 
     bob._objects = {}
 
@@ -210,9 +209,9 @@ def test_obj_not_found(workers):
 
 def test_get_not_permitted(workers):
     bob = workers["bob"]
-    with patch.object(th.Tensor, "allowed_to_get") as mock_allowed_to_get:
+    with patch.object(torch.Tensor, "allowed_to_get") as mock_allowed_to_get:
         mock_allowed_to_get.return_value = False
-        x = th.tensor([1, 2, 3, 4, 5]).send(bob)
+        x = torch.tensor([1, 2, 3, 4, 5]).send(bob)
         with pytest.raises(GetNotPermittedError):
             x.get()
         mock_allowed_to_get.assert_called_once()
@@ -224,7 +223,7 @@ def test_spinup_time(hook):
     spun up inside web frameworks are created quickly enough to not cause timeout errors"""
     data = []
     for i in range(10000):
-        data.append(th.Tensor(5, 5).random_(100))
+        data.append(torch.Tensor(5, 5).random_(100))
     start_time = time()
     dummy = sy.VirtualWorker(hook, id="dummy", data=data)
     end_time = time()
